requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
psycopg2-binary
sqlalchemy
openpyxl
azure-storage-blob
polars
pandas
pyarrow
requests
pytest